+++ src/app/layout.tsx
@@ -1,10 +1,10 @@
 import "./globals.css";
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
-import StoreProvider from "@/components/StoreProvider";
-import ThemeProvider from "@/components/ThemeProvider";
-import ToastProvider from "@/components/ToastProvider";
-import ModalProvider from "@/components/ModalProvider";
+import StoreProvider from "@/components/providers/StoreProvider";
+import ThemeProvider from "@/components/providers/ThemeProvider";
+import ToastProvider from "@/components/providers/ToastProvider";
+import ModalProvider from "@/components/providers/ModalProvider";
 
 export const metadata: Metadata = {
   title: "Chingu Dashboard",
