+++ src/components/providers/ModalProvider.tsx
@@ -2,9 +2,9 @@
 
 import { useEffect, useState } from "react";
 
-import Example1Modal from "./modals/Example1Modal";
-import Example2Modal from "./modals/Example2Modal";
-import FeatureModal from "./modals/FeatureModal";
+import Example1Modal from "@/components/modals/Example1Modal";
+import Example2Modal from "@/components/modals/Example2Modal";
+import FeatureModal from "@/components/modals/FeatureModal";
 
 export default function ModalProvider() {
   const [isMounted, setIsMounted] = useState(false);
